+++ test/streaming/get/test_network_consumer.py
@@ -294,7 +294,7 @@ class TestNetworkConsumer:
         await self.mock_server.validate(client_test, [UnaryGrpc('getNetworkHierarchy', unary_from_fixed(None, _create_hierarchy_response(feeder_network)))])
 
     @pytest.mark.asyncio
-    async def test_get_equipment_container_sends_linked_container_params(self, feeder_network: NetworkService):
+    async def test_get_equipment_container_sends_options(self, feeder_network: NetworkService):
         feeder_mrid = "f001"
 
         async def client_test():
@@ -302,7 +302,8 @@ class TestNetworkConsumer:
                 feeder_mrid,
                 Feeder,
                 IncludedEnergizingContainers.INCLUDE_ENERGIZING_SUBSTATIONS,
-                IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS
+                IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS,
+                NetworkState.ALL_NETWORK_STATE
             )
 
         object_responses = _create_object_responses(feeder_network)
@@ -314,7 +315,8 @@ class TestNetworkConsumer:
                                                 _create_container_responses(
                                                     feeder_network,
                                                     expected_include_energizing_containers=IncludedEnergizingContainers.INCLUDE_ENERGIZING_SUBSTATIONS,
-                                                    expected_include_energized_containers=IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS
+                                                    expected_include_energized_containers=IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS,
+                                                    network_state=NetworkState.ALL_NETWORK_STATE
                                                 )
                                             ]),
                                             StreamGrpc('getIdentifiedObjects', [object_responses, object_responses])
@@ -331,28 +333,14 @@ class TestNetworkConsumer:
         await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers', [_create_container_equipment_responses(feeder_network)])])
 
     @pytest.mark.asyncio
-    async def test_get_equipment_for_container_supports_linked_containers(self, feeder_network: NetworkService):
+    async def test_get_equipment_for_container_sends_options(self, feeder_network: NetworkService):
         async def client_test():
             mor = (
                 await self.client.get_equipment_for_container(
                     "f001",
                     IncludedEnergizingContainers.INCLUDE_ENERGIZING_SUBSTATIONS,
-                    IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS
-                )
-            ).throw_on_error().value
-
-            assert len(mor.objects) == self.service.len_of(Equipment) == 3
-            _assert_contains_mrids(self.service, "fsp", "c2", "tx")
-
-        await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers', [_create_container_equipment_responses(feeder_network)])])
-
-    @pytest.mark.asyncio
-    async def test_get_equipment_for_container_for_all_network_state(self, feeder_network: NetworkService):
-        async def client_test():
-            mor = (
-                await self.client.get_equipment_for_container(
-                    "f001",
-                    network_state=NetworkState.ALL_NETWORK_STATE
+                    IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS,
+                    NetworkState.ALL_NETWORK_STATE
                 )
             ).throw_on_error().value
 
@@ -361,6 +349,8 @@ class TestNetworkConsumer:
 
         await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers',
                                                                  [_create_container_equipment_responses(feeder_network,
+                                                                                                        expected_include_energizing_containers=IncludedEnergizingContainers.INCLUDE_ENERGIZING_SUBSTATIONS,
+                                                                                                        expected_include_energized_containers=IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS,
                                                                                                         network_state=NetworkState.ALL_NETWORK_STATE)])])
 
     @pytest.mark.asyncio
@@ -374,36 +364,23 @@ class TestNetworkConsumer:
         await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers', [_create_container_equipment_responses(feeder_network)])])
 
     @pytest.mark.asyncio
-    async def test_get_equipment_for_containers_supports_linked_containers(self, feeder_network: NetworkService):
+    async def test_get_equipment_for_containers_sends_options(self, feeder_network: NetworkService):
         async def client_test():
             mor = (
                 await self.client.get_equipment_for_containers(
                     ["f001"],
                     IncludedEnergizingContainers.INCLUDE_ENERGIZING_SUBSTATIONS,
-                    IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS
+                    IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS,
+                    NetworkState.ALL_NETWORK_STATE
                 )
             ).throw_on_error().value
 
             assert len(mor.objects) == self.service.len_of(Equipment) == 3
             _assert_contains_mrids(self.service, "fsp", "c2", "tx")
 
-        await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers', [_create_container_equipment_responses(feeder_network)])])
-
-    @pytest.mark.asyncio
-    async def test_get_equipment_for_containers_for_all_network_state(self, feeder_network: NetworkService):
-        async def client_test():
-            mor = (
-                await self.client.get_equipment_for_containers(
-                    ["f001"],
-                    network_state=NetworkState.ALL_NETWORK_STATE
-                )
-            ).throw_on_error().value
-
-            assert len(mor.objects) == self.service.len_of(Equipment) == 3
-            _assert_contains_mrids(self.service, "fsp", "c2", "tx")
-
-        await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers',
-                                                                 [_create_container_equipment_responses(feeder_network,
+        await self.mock_server.validate(client_test, [StreamGrpc('getEquipmentForContainers', [_create_container_equipment_responses(feeder_network,
+                                                                                                        expected_include_energizing_containers=IncludedEnergizingContainers.INCLUDE_ENERGIZING_SUBSTATIONS,
+                                                                                                        expected_include_energized_containers=IncludedEnergizedContainers.INCLUDE_ENERGIZED_LV_FEEDERS,
                                                                                                         network_state=NetworkState.ALL_NETWORK_STATE)])])
 
     @pytest.mark.asyncio
@@ -594,12 +571,19 @@ def _to_network_identified_object(obj) -> NetworkIdentifiedObject:
 
 
 def _create_container_equipment_responses(ns: NetworkService, mrids: Optional[Iterable[str]] = None,
-                                          network_state: NetworkState = NetworkState.NORMAL_NETWORK_STATE) \
+                                          expected_include_energizing_containers: Optional[int] = None,
+                                          expected_include_energized_containers: Optional[int] = None,
+                                          network_state: NetworkState = None) \
     -> Callable[[GetEquipmentForContainersRequest], Generator[GetEquipmentForContainersResponse, None, None]]:
     valid: Dict[str, EquipmentContainer] = {mrid: ns[mrid] for mrid in mrids} if mrids else ns
 
     def responses(request: GetEquipmentForContainersRequest):
-        assert request.networkState == network_state
+        if expected_include_energizing_containers is not None:
+            assert request.includeEnergizingContainers == expected_include_energizing_containers
+        if expected_include_energized_containers is not None:
+            assert request.includeEnergizedContainers == expected_include_energized_containers
+        if network_state is not None:
+            assert request.networkState == network_state
         for mrid in request.mrids:
             ec = valid[mrid]
             if ec:
@@ -687,7 +671,8 @@ def _create_container_responses(
     ns: NetworkService,
     mrids: Optional[Iterable[str]] = None,
     expected_include_energizing_containers: Optional[int] = None,
-    expected_include_energized_containers: Optional[int] = None
+    expected_include_energized_containers: Optional[int] = None,
+    network_state: Optional[int] = None
 ) -> Callable[[GetEquipmentForContainersRequest], Generator[GetEquipmentForContainersResponse, None, None]]:
     valid: Dict[str, EquipmentContainer] = {mrid: ns[mrid] for mrid in mrids} if mrids else ns
 
@@ -696,6 +681,8 @@ def _create_container_responses(
             assert request.includeEnergizingContainers == expected_include_energizing_containers
         if expected_include_energized_containers is not None:
             assert request.includeEnergizedContainers == expected_include_energized_containers
+        if network_state is not None:
+            assert request.networkState == network_state
 
         for mrid in request.mrids:
             container = valid[mrid]